create sa separate module for erosion calculation
